feat: final round of refactoring, readme overhaul 🎓
